Fix build, cleanup.
