--- orttraining/orttraining/core/optimizer/bias_dropout_fusion.h
@@ -1,24 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-#pragma once
-
-#include "core/optimizer/graph_transformer.h"
-
-namespace onnxruntime {
-
-/**
-@Class BiasDropoutFusion
-
-Fuse Add + Dropout + optional Add to BiasDropoutFusion
-
-*/
-class BiasDropoutFusion : public GraphTransformer {
- public:
-  BiasDropoutFusion(const std::unordered_set<std::string>& compatible_execution_providers = {}) noexcept
-      : GraphTransformer("BiasDropoutFusion", compatible_execution_providers) {}
-
-  Status ApplyImpl(Graph& graph, bool& modified, int graph_level, const logging::Logger& logger) const override;
-};
-
-}  // namespace onnxruntime
