+++ winml/api/Microsoft.AI.MachineLearning.Experimental.idl
@@ -20,7 +20,6 @@ import "Windows.AI.MachineLearning.idl";
 #define ROOT_NS Microsoft
 #endif
 
-
 namespace ROOT_NS.AI.MachineLearning.Experimental {
   runtimeclass LearningModelBuilder;
